test: update and fix youtube tests
